requirements.txt
matplotlib==3.10.7
mysql_connector_repackaged==0.3.1
texttable==1.7.0
